+++ src/state/history.ts
@@ -17,6 +17,10 @@ export class StateHistory {
         return StateKeeper.HISTORY;
     }
 
+    get storeHistoryItems() {
+        return this.options.storeHistoryItems;
+    }
+
     init(initialState: any) {
         StateHistory.initialState = initialState;
     }
@@ -25,9 +29,11 @@ export class StateHistory {
         this.options = { ...this.options, ...options };
     }
 
-    add(state: any) {
+    setCurrentState(state: any) {
         StateKeeper.CURRENT_STATE = state;
+    }
 
+    add(item: HistoryItem) {
         if (!this.options.collectHistory) {
             return;
         }
@@ -36,7 +42,7 @@ export class StateHistory {
             StateKeeper.HISTORY.shift();
         }
 
-        StateKeeper.HISTORY.push(state.toJS());
+        StateKeeper.HISTORY.push(item);
     }
 }
 
@@ -48,4 +54,9 @@ export class StateKeeper {
 export interface StateHistoryOptions {
     collectHistory?: boolean;
     storeHistoryItems?: number | null;
+}
+
+export interface HistoryItem {
+    message: string;
+    state: any;
 }